+++ modules/components/widgets/index.js
@@ -1,5 +1,6 @@
 
 export {default as ValueFieldWidget} from '../ValueField';
 export {default as FuncWidget} from '../FuncWidget';
-export * from './antd/index.js';
+
+//export * from './antd/index.js';
 export * from './vanilla/index.js';
