+++ package.json
@@ -41,7 +41,6 @@
   },
   "homepage": "https://github.com/ukrbublik/react-awesome-query-builder",
   "dependencies": {
-    "antd": "^3.25.0",
     "classnames": "^2.2.5",
     "clone": "^2.1.1",
     "immutable": "^3.7.6",
@@ -55,6 +54,7 @@
     "transit-js": "^0.8.846"
   },
   "peerDependencies": {
+    "antd": "^3.25.0",
     "react": "^0.14.6 || ^15.0.0 || ^16.4.1 ",
     "react-dom": "^0.14.6 || ^15.0.0 || ^16.4.1",
     "react-redux": "^4.0.6 || ^5.0.0",
